+++ ChangeLog.md
@@ -26,6 +26,7 @@
 - [KT-11638](https://youtrack.jetbrains.com/issue/KT-11638) Fixed hashCode() implementation in "Generate equals/hashCode" action
 - Configure Kotlin: show only changed files in the notification "Kotlin not configured"
 - Configure Kotlin: restore all changed files in undo action
+- Convert java expression to kotlin when open Evaluate Expression dialog from Variables View
 
 ### Java to Kotlin converter